requirements.txt
GoogleNews == 1.5.9
googlesearch-python == 1.0.1      
beautifulsoup4 == 4.9.3
numpy== 1.21.1
pandas==1.1.5
googletrans>=3.1.0a0
Wikipedia-API>=0.5.4
sentence_transformers
